Fix Python 2.x compatibility

+++ vc2_conformance/picture_generators.py
@@ -491,7 +491,7 @@ def moving_sprite(video_parameters, picture_coding_mode, duration=1.0):
     # Generate frames
     num_samples, relative_rate = seconds_to_samples(video_parameters, duration)
     x_step_size = 16 // relative_rate
-    for px in np.arange(num_samples) * x_step_size:
+    for px in np.arange(num_samples, dtype=int) * x_step_size:
         # For bizarre, tiny picture formats too small for the sprite, clip the
         # sprite to fit
         sw = min(frame_width, sprite_width)
